requirements.txt
torch~=2.0.1
matplotlib~=3.7.2
numpy~=1.25.1
wandb~=0.15.5
torchvision~=0.15.2
pandas~=2.0.3
pillow~=10.0.0